fixup! feat: Preserve user-managed fields when creating namespace
Fix the test entrypoint name, and move integration tests to their own file
