+++ pkg/handlers/generic/lifecycle/utils/utils_test.go
@@ -6,15 +6,10 @@ package utils
 import (
 	"testing"
 
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	"github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/test/helpers"
 )
 
 func TestCreateConfigMapForCRS(t *testing.T) {
@@ -97,51 +92,3 @@ provisioner: ""`,
 		})
 	}
 }
-
-var _ = Describe("Namespace", func() {
-	It("creates a new namespace", func(ctx SpecContext) {
-		c, err := helpers.TestEnv.GetK8sClient()
-		Expect(err).To(BeNil())
-
-		namespaceName := "new"
-
-		Expect(EnsureNamespace(ctx, c, namespaceName)).To(Succeed())
-		Expect(c.Delete(ctx, &corev1.Namespace{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: namespaceName,
-			},
-		})).To((Succeed()))
-	})
-
-	It("updates a namespace, preserving user-managed fields", func(ctx SpecContext) {
-		c, err := helpers.TestEnv.GetK8sClient()
-		Expect(err).To(BeNil())
-
-		namespaceName := "existing"
-		Expect(c.Create(ctx,
-			&corev1.Namespace{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: namespaceName,
-					Labels: map[string]string{
-						"userkey": "uservalue",
-					},
-				},
-			})).To(Succeed())
-
-		Expect(EnsureNamespace(ctx, c, namespaceName)).To(Succeed())
-
-		ns := &corev1.Namespace{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: namespaceName,
-			},
-		}
-		Expect(c.Get(ctx, client.ObjectKeyFromObject(ns), ns)).To(Succeed())
-		Expect(ns.GetLabels()).To(HaveKeyWithValue("userkey", "uservalue"))
-
-		Expect(c.Delete(ctx, &corev1.Namespace{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: namespaceName,
-			},
-		})).To((Succeed()))
-	})
-})
